chore: refactor logic into initialize_contracts
