+++ bouncer/commands/setup_vaults.ts
@@ -17,18 +17,19 @@ import {
   sleep,
   handleSubstrateError,
   getEvmEndpoint,
-  getContractAddress,
   getSolConnection,
 } from '../shared/utils';
 import { aliceKeyringPair } from '../shared/polkadot_keyring';
-import { getKeyManagerAbi } from '../shared/contract_interfaces';
-import { signAndSendTxEvm } from '../shared/send_evm';
-import { initializeSolanaPrograms } from '../shared/initialize_solana_programs';
+import {
+  initializeArbitrumContracts,
+  initializeSolanaPrograms,
+} from '../shared/initialize_contracts';
 
 async function main(): Promise<void> {
   const btcClient = getBtcClient();
   const arbClient = new Web3(getEvmEndpoint('Arbitrum'));
   const alice = await aliceKeyringPair();
+  const solClient = getSolConnection();
 
   const chainflip = await getChainflipApi();
   const polkadot = await getPolkadotApi();
@@ -139,35 +140,11 @@ async function main(): Promise<void> {
 
   // Step 6
   console.log('Inserting Arbitrum key in the contracts');
-  const keyManagerAddress = getContractAddress('Arbitrum', 'KEY_MANAGER');
+  await initializeArbitrumContracts(arbClient, arbKey);
 
-  const keyManagerContract = new arbClient.eth.Contract(
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (await getKeyManagerAbi()) as any,
-    keyManagerAddress,
-  );
-  const txData = keyManagerContract.methods
-    .setAggKeyWithGovKey({
-      pubKeyX: arbKey.pubKeyX,
-      pubKeyYParity: arbKey.pubKeyYParity === 'Odd' ? 1 : 0,
-    })
-    .encodeABI();
-  await signAndSendTxEvm('Arbitrum', keyManagerAddress, '0', txData);
-
-  // Temporal workaround if running the bouncer without the Solana node
-  let solanaIsRunning = false;
-  try {
-    const solClient = getSolConnection();
-    await solClient.getGenesisHash();
-    solanaIsRunning = true;
-  } catch (e) {
-    console.log('Solana not running, skipping key initialization');
-  }
-  if (solanaIsRunning) {
-    // Using arbitrary key for now, we will use solKey generated by SC
-    const solKey = '0x25fcb03ab6435d106b5df1e677f3c6a10a7b22719deedeb3761c005e1306423d';
-    await initializeSolanaPrograms(solKey);
-  }
+  // Using arbitrary key for now, we will use solKey generated by SC
+  const solKey = '0x25fcb03ab6435d106b5df1e677f3c6a10a7b22719deedeb3761c005e1306423d';
+  await initializeSolanaPrograms(solClient, solKey);
 
   // Step 7
   console.log('Registering Vaults with state chain');
